Move ensurePublicKeyEncoding to assertions & use keyType.
